what_version_are_you {package.e}	R Documentation

Just some example function. But this function will also be present in another package.

Description

Just some example function. But this function will also be present in another package.

Usage

what_version_are_you()

[Package package.e version 1.7.0 Index]